refactor: split WelcomeByeGuild logic into multiple files
